Tests updated.

+++ lionengine-core-awt-test/src/test/java/com/b3dgs/lionengine/core/awt/UtilityMediaTest.java
@@ -149,7 +149,10 @@ public class UtilityMediaTest
         }
         finally
         {
-            UtilFile.deleteFile(media.getFile());
+            if (media.exists())
+            {
+                UtilFile.deleteFile(media.getFile());
+            }
         }
     }
